refactor(atomic): decouple request transformers from mock classes
Move RequestTransformer type and skipOnError utility to a dedicated
_request-transformer.ts module. Replace single withRequestTransformer()
with addRequestTransformer()/removeRequestTransformer() supporting
multiple transformers per endpoint.

Remove enableInteractiveFacets() methods from MockSearchApi and
MockCommerceApi to avoid tight coupling between mock nexus files
and transformer modules. Stories now import and register transformers
directly, so Chromatic TurboSnap only re-tests stories that actually
use a modified transformer.
