+++ packages/atomic/storybook-utils/api/_request-transformer.ts
@@ -1,0 +1,4 @@
+export type RequestTransformer<TResponse> = (
+  body: unknown,
+  response: TResponse
+) => TResponse;
